feat: #190 🎸 不要なコンポーネント削除
