--- components/UserHeader.vue
@@ -1,132 +1,0 @@
-<script setup lang="ts">
-import { useRouter } from '#app'
-import { mdiAccount, mdiSwordCross, mdiHome, mdiAccountCircle } from '@mdi/js'
-import { ref, useUser } from '#imports'
-
-type Link = {
-  title: string
-  icon: string
-  url: string
-}
-
-const { loginedUser, logout } = useUser()
-const router = useRouter()
-
-const isDrawerOpen = ref<boolean>(false)
-const linkList: Link[] = [
-  {
-    title: 'プロフィール変更',
-    icon: mdiAccount,
-    url: '/profile',
-  },
-  {
-    title: 'ゲーム履歴',
-    icon: mdiSwordCross,
-    url: '/gameHistory',
-  },
-  {
-    title: 'ホーム',
-    icon: mdiHome,
-    url: '/',
-  },
-]
-
-const handleLogout = async () => {
-  const { success } = await logout()
-  if (success) {
-    router.push('/login')
-  }
-}
-</script>
-
-<template>
-  <div>
-    <v-navigation-drawer v-model="isDrawerOpen" app clipped>
-      <v-list>
-        <v-list-item v-if="loginedUser" :title="loginedUser.name" class="my-4">
-          <template #prepend>
-            <v-avatar
-              v-if="loginedUser && loginedUser.iconImageUrl"
-              :image="loginedUser.iconImageUrl"
-              class="user-icon"
-            />
-            <v-icon
-              v-else
-              color="grey"
-              class="user-icon"
-              :icon="mdiAccountCircle"
-            />
-          </template>
-        </v-list-item>
-      </v-list>
-
-      <v-divider />
-
-      <v-list dense>
-        <v-list-item v-for="link in linkList" :key="link.title" :to="link.url">
-          <template #prepend>
-            <v-icon :icon="link.icon" />
-          </template>
-          <v-list-item-title class="text-subtitle-2">{{
-            link.title
-          }}</v-list-item-title>
-        </v-list-item>
-      </v-list>
-
-      <template #append>
-        <div class="pa-2">
-          <v-btn
-            v-if="loginedUser"
-            block
-            variant="outlined"
-            rounded="xl"
-            color="forest-shade"
-            @click="handleLogout"
-            >ログアウト</v-btn
-          >
-        </div>
-      </template>
-    </v-navigation-drawer>
-    <v-app-bar flat app>
-      <template #prepend>
-        <v-app-bar-nav-icon @click="isDrawerOpen = !isDrawerOpen" />
-      </template>
-      <nuxt-link to="/profile">
-        <v-avatar
-          v-if="loginedUser && loginedUser.iconImageUrl"
-          :image="loginedUser.iconImageUrl"
-          class="user-icon"
-        />
-        <v-icon
-          v-else
-          color="grey"
-          class="user-icon"
-          :icon="mdiAccountCircle"
-        />
-      </nuxt-link>
-
-      <v-app-bar-title v-if="loginedUser"
-        >{{ loginedUser.name }}
-        <div>
-          <span class="star"> ★ </span>
-          ×{{ loginedUser.stars }}
-        </div>
-      </v-app-bar-title>
-    </v-app-bar>
-  </div>
-</template>
-
-<style scoped>
-.v-app-bar {
-  border-bottom: 1px solid grey;
-}
-
-.user-icon {
-  width: 15vw;
-  max-width: 48px;
-  height: 15vw;
-  max-height: 48px;
-  border-radius: 50%;
-  background-color: white;
-}
-</style>
